+++ src/MenuItem.jsx
@@ -1,5 +1,3 @@
-
-
 import React from 'react';
 import {joinClasses, classSet, KeyCode} from 'rc-util';
 
@@ -38,15 +36,20 @@ class MenuItem extends React.Component {
   onClick(e) {
     const props = this.props;
     const eventKey = props.eventKey;
-    props.onClick(eventKey, this, e);
+    const info = {
+      key: eventKey,
+      item: this,
+      domEvent: e,
+    };
+    props.onClick(info);
     if (props.multiple) {
       if (props.selected) {
-        props.onDeselect(eventKey, this, e);
+        props.onDeselect(info);
       } else {
-        props.onSelect(eventKey, this, e);
+        props.onSelect(info);
       }
     } else if (!props.selected) {
-      props.onSelect(eventKey, this, e);
+      props.onSelect(info);
     }
   }
 
@@ -88,11 +91,15 @@ class MenuItem extends React.Component {
         onMouseEnter: this.onMouseEnter,
       };
     }
+    const style = {};
+    if (props.mode === 'inline') {
+      style.paddingLeft = props.inlineIndent * props.level;
+    }
     return (
-      <li
+      <li style={style}
         {...attrs}
         {...mouseEvent}>
-      {props.children}
+        {props.children}
       </li>
     );
   }